backend/requirements.txt
Flask>=2.2
flask-cors>=3.0
# pyswisseph version pinned to a known available release compatible with build
pyswisseph==2.10.3.2
